Test23
Toutes les fonctions
la recup des variables par l'interface
pas encore les calculs pour afficher resultats
pas encore de camenbert

+++ JavaApplication7/src/javaapplication7/Panneau.java
@@ -72,7 +72,7 @@ public class Panneau {
         this.rendementp = rendementp;
     }
 ///////////////////////////////////////////////
-    public int getSurfacep() {
+    public double getSurfacep() {
         return surfacep;
     }
